feat: LottoTicket 클래스 -> 인터페이스로 변경 AutoLottoTicket 클래스 생성 ManualLottoTicket 클래스 생성
