+++ src/main/java/step5/LottoApp.java
@@ -1,11 +1,13 @@
 package step5;
 
+import step5.domain.LottoTicket;
+import step5.domain.ManualLottoTicket;
 import step5.view.InputView;
 
 public class LottoApp {
     public static void main(String[] args) {
         Integer purchaseAmount = InputView.requirePurchaseAmount();
         Integer numberOfManualLotto = InputView.requireNumberOfManualLotto();
-        String manualLottoNumber = InputView.requireManualLottoNumber();
+        LottoTicket manualLottoTicket = new ManualLottoTicket(InputView.requireManualLottoNumber());
     }
 }
